The print helper takes a word, so the cheat buffer's index can too
io_print_dos and io_cp437_utf8 took their length and capacity as
uint32_t, and that was the only reason the cheat screen's index had to be
one - it is handed straight to them. Both write into buffers of 1024 and
256, so a word is the width, and the caller narrows with them.

That also settles the earlier objection to narrowing it at all: as a
uint8_t the `n < sizeof line` guarding a 256-byte buffer can never be
false, which is why it stayed wide. A uint16_t can exceed 256, so the
check still means something.

game.c now holds no uint32_t at all.

18 routines agree with the original on the menu route, and 1,500 frames
on panel 8 are identical throughout.

+++ reconstruct/src/game.c
@@ -7692,11 +7692,11 @@ int32_t cheat_sequence(char key)
     {
         uint8_t line[256];
         /* `n` indexes a buffer this port owns rather than a register the
-         * original has, and it stays wide on purpose: as a uint8_t the
-         * `n < sizeof line` below can never be false, and a bounds check
-         * that cannot fail is one that stops protecting the moment the
-         * buffer shrinks. */
-        uint32_t n = 0;
+         * original has. A uint8_t would make the `n < sizeof line` below
+         * unfalsifiable - the buffer is 256 - and a bounds check that cannot
+         * fail stops protecting the moment the buffer shrinks. A word can
+         * still be too big for it, which is the point of having it. */
+        uint16_t n = 0;
         uint8_t ah = ' ';
         const uint8_t *s = runtime.cheat_text;
         for (;;) {
